refactored xml import so it should work, but still needs modifications to properly handle validation/files/tracked jobs

+++ src/ODR/AdminBundle/Command/XMLImportFileDownloadCommand.php
@@ -2,15 +2,15 @@
 
 /**
 * Open Data Repository Data Publisher
-* ImportWorker Command
+* XMLImportFileDownload Command
 * (C) 2015 by Nathan Stone (nate.stone@opendatarepository.org)
 * (C) 2015 by Alex Pires (ajpires@email.arizona.edu)
 * Released under the GPLv2
 *
 * This Symfony console command takes beanstalk jobs from the
-* import_worker tube and passes the parameters to XMLController,
-* which will attempt to import the contents of an XML file into
-* a DataRecord entity in the database.
+* import_file tube and passes the parameters to XMLController,
+* which will attempt to download the described File or Image
+* entity from some remote server as part of an XML import.
 *
 */
 
@@ -29,15 +29,15 @@ use Symfony\Component\Console\Output\OutputInterface;
 use ODR\AdminBundle\Entity\DataRecord;
 use ODR\AdminBundle\Entity\DataType;
 
-class ImportWorkerCommand extends ContainerAwareCommand
+class XMLImportFileDownloadCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('odr_import:worker')
-            ->setDescription('Attempts to import an XML file into the database...');
+            ->setName('odr_xml_import:file_download')
+            ->setDescription('Attempts to download a file from a remote server for importing purposes...');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -45,33 +45,25 @@ class ImportWorkerCommand extends ContainerAwareCommand
         // Only need to load these once...
         $container = $this->getContainer();
         $logger = $container->get('logger');
-        $router = $container->get('router');
         $pheanstalk = $container->get('pheanstalk');
 
         // Run command until manually stopped
         while (true) {
             $job = null;
             try {
-                // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_import_datarecord')->ignore('default')->reserve();
-                $job = $pheanstalk->watch('import_datarecord')->ignore('default')->reserve();
+                // Wait for a job
+                $job = $pheanstalk->watch('import_file')->ignore('default')->reserve();
 
                 // Get Job Data
                 $data = json_decode($job->getData());
 
                 // 
-                $str = '';
-                if ($data->datatype_id !== '')
-                    $str = 'ImportWorkerCommand.php: Attempting to import "'.$data->xml_filename.'" for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
-                else if ($data->datafield_id !== '')
-                    $str = 'ImportWorkerCommand.php: Attempting to import "'.$data->xml_filename.'" for DataField '.$data->datafield_id.' from '.$data->memcached_prefix.'...';
-                else
-                    throw new \Exception('Invalid job data');
+                $str = 'XMLImportFileDownloadCommand.php: Attempting to import file at "'.$data->href.'" into '.$data->object_type.' '.$data->object_id.' from '.$data->memcached_prefix.'...';
 
                 $current_time = new \DateTime();
                 $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );
                 $output->writeln($str);
-                $logger->info('ImportWorkerCommand.php: '.$str);
+                $logger->info('XMLImportFileDownloadCommand.php: '.$str);
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
@@ -79,7 +71,7 @@ class ImportWorkerCommand extends ContainerAwareCommand
 $output->writeln($data->url);
 
                 // Create the required url and the parameters to send
-                $parameters = array('datatype_id' => $data->datatype_id, 'datafield_id' => $data->datafield_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id, 'xml_filename' => $data->xml_filename);
+                $parameters = array('href' => $data->href, 'object_type' => $data->object_type, 'object_id' => $data->object_id, 'user_id' => $data->user_id, 'api_key' => $data->api_key);
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -89,7 +81,7 @@ $output->writeln($data->url);
                         CURLOPT_FRESH_CONNECT => 1,
                         CURLOPT_RETURNTRANSFER => 1,
                         CURLOPT_FORBID_REUSE => 1,
-                        CURLOPT_TIMEOUT => 300,
+                        CURLOPT_TIMEOUT => 120,
                         CURLOPT_POSTFIELDS => http_build_query($parameters)
                     )
                 );
@@ -102,16 +94,22 @@ $output->writeln($data->url);
                 // Do things with the response returned by the controller?
                 $result = json_decode($ret);
                 if ( isset($result->r) && isset($result->d) ) {
-                    if ( $result->r == 0 )
+                    if ( $result->r == 0 ) {
                         $output->writeln( $result->d );
-                    else
+                    }
+                    else if ( $result->r == 6 ) {
+                        // Could not resolve host error
+                        throw new \Exception( 'retry' );
+                    }
+                    else {
                         throw new \Exception( $result->d );
+                    }
                 }
                 else {
                     // Should always be a json return...
                     throw new \Exception( print_r($ret, true) );
                 }
-//$logger->debug('ImportWorkerCommand.php: curl results...'.print_r($result, true));
+//$logger->debug('XMLImportFileDownloadCommand.php: curl results...'.print_r($result, true));
 
                 // Done with this cURL object
                 curl_close($ch);
@@ -121,11 +119,20 @@ $output->writeln($data->url);
 
             }
             catch (\Exception $e) {
-                $output->writeln($e->getMessage());
-                $logger->err('ImportWorkerCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('XMLImportFileDownloadCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue doesn't hang, in theory
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+                }
+                else {
+                    $output->writeln($e->getMessage());
+                    $logger->err('XMLImportFileDownloadCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
